moved some of the HDF5 constants into the connectors.hdf5 module
